Added reports section. Cashbox balance and profit moved to reports section
