+++ app/src/main/res/values-kaa/strings.xml
@@ -146,4 +146,5 @@
     <string name="profit_title">Sap payda</string>
     <string name="err_valid_sum">Haqıyqıy summa kiritiń</string>
     <string name="warning">Isenimińiz kámil me?</string>
+    <string name="reports">Esabatlar</string>
 </resources>
