+++ app/src/main/res/values-uz/strings.xml
@@ -146,4 +146,5 @@
     <string name="profit_title">Sof foyda</string>
     <string name="err_valid_sum">Haqiqiy summa kiriting</string>
     <string name="warning">Ishonchingiz komilmi?</string>
+    <string name="reports">Hisobotlar</string>
 </resources>
