+++ app/src/main/res/values/strings.xml
@@ -180,4 +180,5 @@
     <string name="profit_title">Sof foyda</string>
     <string name="err_valid_sum">Haqiqiy summa kiriting</string>
     <string name="warning">Ishonchingiz komilmi?</string>
+    <string name="reports">Hisobotlar</string>
 </resources>
